+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+* Refactored `BTLxPart.shape_strings` to use `Brep.to_polygons()` instead of raw Brep face edge traversal for more robust and accurate shape geometry generation.
 
 ### Removed